+++ src/actions/actions.js
@@ -33,12 +33,12 @@ module.exports = (router) => {
      */
     actions: hook.alter('actions', {
       email: require('./EmailAction')(router),
-      webhook: require('./WebhookAction')(router),
-      sql: require('./SQLAction')(router),
-      role: require('./RoleAction')(router),
+      login: require('./LoginAction')(router),
       resetpass: require('./ResetPassword')(router),
+      role: require('./RoleAction')(router),
       save: require('./SaveSubmission')(router),
-      login: require('./LoginAction')(router)
+      sql: require('./SQLAction')(router),
+      webhook: require('./WebhookAction')(router),
     }),
 
     /**
@@ -179,7 +179,6 @@ module.exports = (router) => {
 
         async.eachSeries(actions, (action, cb) => {
           this.shouldExecute(action, req).then(execute => {
-            debug.action(`execute (${execute}):`, action);
             if (!execute) {
               return cb();
             }
@@ -204,25 +203,22 @@ module.exports = (router) => {
                 }
               ]
             }, req), (err, actionItem) => {
+              // Mongoose has issues if you call "save" too frequently on the same item. We need to wait till the previous
+              // save is complete before calling again.
+              let lastSavePromise = Promise.resolve();
               const setActionItemMessage = (message, data = {}, state = null) => {
-                actionItem.messages.push({
-                  datetime: new Date(),
-                  info: message,
-                  data
-                });
+                lastSavePromise.then(() => {
+                  actionItem.messages.push({
+                    datetime: new Date(),
+                    info: message,
+                    data
+                  });
 
-                if (state) {
-                  actionItem.state = state;
-                }
-
-                // Update the action item.
-                router.formio.mongoose.models.actionItem.updateOne({
-                  _id: actionItem._id
-                }, {
-                  $set: {
-                    state: actionItem.state,
-                    messages: actionItem.messages
+                  if (state) {
+                    actionItem.state = state;
                   }
+
+                  lastSavePromise = actionItem.save();
                 });
               };