+++ src/middleware/tokenHandler.js
@@ -1,10 +1,11 @@
 'use strict';
 
 const jwt = require('jsonwebtoken');
+const _ = require('lodash');
 const util = require('../util/util');
 const debug = {
   error: require('debug')('formio:error'),
-  handler: require('debug')('formio:middleware:tokenHandler')
+  handler: require('debug')('formio:middleware:tokenHandler'),
 };
 
 /**
@@ -18,10 +19,13 @@ const debug = {
  * @returns {Function}
  *   The middleware for an Express endpoint.
  */
-module.exports = function(router) {
+module.exports = (router) => {
   // Load the form.io hooks.
   const hook = require('../util/hook')(router.formio);
   const formioCache = require('../cache/cache')(router);
+  const {
+    jwt: jwtConfig,
+  } = router.formio.config;
 
   /**
    * Util function to update the jwt in the response.
@@ -55,7 +59,7 @@ module.exports = function(router) {
    * @param next
    */
   const userHandler = (req, res, decoded, token, user, next) => {
-    hook.alter('user', user, function(err, user) {
+    hook.alter('user', user, (err, user) => {
       if (err) {
         return next();
       }
@@ -67,7 +71,7 @@ module.exports = function(router) {
       req.token = decoded;
 
       // Refresh the token that is sent back to the user when appropriate.
-      req.tokenIssued = parseInt(Date.now() / 1000);
+      req.tokenIssued = Math.trunc(Date.now() / 1000);
       generateToken(token, decoded, res, req);
       router.formio.log('Token', req, 'Using normal token');
       if (req.user) {
@@ -77,7 +81,7 @@ module.exports = function(router) {
     });
   };
 
-  return function tokenHandler(req, res, next) {
+  return (req, res, next) => {
     /* eslint-disable max-statements */
     // If someone else provided then skip.
     if (req.user && req.token && res.token) {
@@ -85,7 +89,7 @@ module.exports = function(router) {
     }
 
     const token = util.getRequestValue(req, 'x-jwt-token');
-    const noToken = function() {
+    const noToken = () => {
       router.formio.log('Token', req, 'No token found');
       // Try the request with no tokens.
       delete req.headers['x-jwt-token'];
@@ -96,12 +100,13 @@ module.exports = function(router) {
       return next();
     };
 
-    if (req.headers['x-token']) {
-      const keys = process.env.API_KEYS
-        ? process.env.API_KEYS.split(',').map(key => key.trim()).filter(key => !!key)
+    const apiKey = req.headers['x-token'];
+    if (apiKey) {
+      const apiKeys = process.env.API_KEYS
+        ? process.env.API_KEYS.split(',').map((key) => key.trim()).filter(Boolean)
         : [];
 
-      if (keys.includes(req.headers['x-token'])) {
+      if (apiKeys.includes(apiKey)) {
         router.formio.audit('AUTH_APIKEY', req);
         req.isAdmin = true;
         req.permissionsChecked = true;
@@ -117,7 +122,7 @@ module.exports = function(router) {
     }
 
     // Decode/refresh the token and store for later middleware.
-    jwt.verify(token, router.formio.config.jwt.secret, function(err, decoded) {
+    jwt.verify(token, jwtConfig.secret, (err, decoded) => {
       if (err || !decoded) {
         debug.handler(err || `Token could not decoded: ${token}`);
         router.formio.audit('EAUTH_TOKENBAD', req, err);
@@ -138,86 +143,90 @@ module.exports = function(router) {
         }
       }
 
-      // Check to see if this token is allowed to access this path.
-      if (!router.formio.auth.isTokenAllowed(req, decoded)) {
-        router.formio.audit('EAUTH_TEMPTOKEN_ALLOW', req, decoded.allow);
-        return noToken();
-      }
-
-      // If this is a temporary token, then decode it and set it in the request.
-      if (decoded.temp) {
-        router.formio.audit('AUTH_TEMPTOKEN', req);
-        router.formio.log('Token', req, 'Using temp token');
-        debug.handler('Temp token');
-        req.tempToken = decoded;
-        req.user = null;
-        req.token = null;
-        res.token = null;
-        return next();
-      }
-
-      // Allow external tokens.
-      if (!hook.alter('external', decoded, req)) {
-        router.formio.audit('AUTH_EXTERNALTOKEN', req, decoded.allow);
-        decoded.user.project = decoded.project._id;
-        return userHandler(req, res, decoded, token, decoded.user, next);
-      }
-
-      // See if this is a remote token.
-      if (decoded.project && decoded.permission) {
-        router.formio.audit('AUTH_REMOTETOKEN', req, decoded.origin);
-        req.userProject = decoded.project;
-        req.remotePermission = decoded.permission;
-        return userHandler(req, res, decoded, token, decoded.user, next);
-      }
-
-      router.formio.audit('AUTH_TOKEN', {
-        ...req,
-        userId: decoded.user._id,
-      });
-      if (decoded.isAdmin) {
-        router.formio.log('Token', req, 'User is admin');
-        if (req.user) {
-          router.formio.log('User', req, req.user._id);
+      hook.alter('tokenDecode', decoded, req, (err, decoded) => {
+        // Check to see if this token is allowed to access this path.
+        if (!router.formio.auth.isTokenAllowed(req, decoded)) {
+          return noToken();
         }
-        req.permissionsChecked = true;
-        req.isAdmin = true;
-        req.token = decoded;
-        return next();
-      }
 
-      if (!decoded.form || !decoded.form._id) {
-        return noToken();
-      }
-      if (!decoded.user || !decoded.user._id) {
-        return noToken();
-      }
-
-      // Load the user submission.
-      const cache = router.formio.cache || formioCache;
-      cache.loadSubmission(req, decoded.form._id, decoded.user._id, function(err, user) {
-        if (err) {
-          // Couldn't load the use, try to fail safely.
-          user = decoded.user;
-        }
-        else if (!user) {
+        // If this is a temporary token, then decode it and set it in the request.
+        if (decoded.temp) {
+          router.formio.log('Token', req, 'Using temp token');
+          debug.handler('Temp token');
+          req.tempToken = decoded;
           req.user = null;
           req.token = null;
           res.token = null;
           return next();
         }
 
-        // Allow anyone to alter the user.
-        debug.handler(user);
-
-        // Check if the user has reset the password since the token was issued.
-        if (user.metadata && user.metadata.jwtIssuedAfter && decoded.iat < user.metadata.jwtIssuedAfter) {
-          router.formio.log('Token', req, 'Token No Longer Valid');
-          return res.status(440).send('Token No Longer Valid');
+        // Allow external tokens.
+        if (!hook.alter('external', decoded, req)) {
+          decoded.user.project = decoded.project._id;
+          return userHandler(req, res, decoded, token, decoded.user, next);
         }
 
-        // Call the user handler.
-        userHandler(req, res, decoded, token, user, next);
+        // See if this is a remote token.
+        if (decoded.project && decoded.permission) {
+          req.userProject = decoded.project;
+          req.remotePermission = decoded.permission;
+          return userHandler(req, res, decoded, token, decoded.user, next);
+        }
+
+        if (decoded.isAdmin) {
+          router.formio.log('Token', req, 'User is admin');
+          if (req.user) {
+            router.formio.log('User', req, req.user._id);
+          }
+          req.permissionsChecked = true;
+          req.isAdmin = true;
+          req.token = decoded;
+          return next();
+        }
+
+        const formId = _.get(decoded, 'form._id');
+        const userId = _.get(decoded, 'user._id');
+
+        if (!formId || !userId) {
+          return noToken();
+        }
+
+        // Load the user submission.
+        const cache = router.formio.cache || formioCache;
+        cache.loadSubmission(req, formId, userId, (err, user) => {
+          if (err) {
+            // Couldn't load the user, try to fail safely.
+            user = decoded.user;
+          }
+          else if (!user) {
+            req.user = null;
+            req.token = null;
+            res.token = null;
+            return next();
+          }
+
+          debug.handler(user);
+
+          hook.alter('validateToken', req, decoded, user, (err) => {
+            if (err) {
+              return res.status(440).send(err);
+            }
+
+            // Check if the user has reset the password since the token was issued.
+            if (
+              !req.skipTokensValidation
+              && user.metadata
+              && user.metadata.jwtIssuedAfter
+              && decoded.iat < user.metadata.jwtIssuedAfter
+            ) {
+              router.formio.log('Token', req, 'Token No Longer Valid');
+              return res.status(440).send('Token No Longer Valid');
+            }
+
+            // Call the user handler.
+            userHandler(req, res, decoded, token, user, next);
+          });
+        });
       });
     });
   };